+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This file will document the most important changes for each released version.
 
 ## [v0.1.4]
 - Logs now go to stderr, leaving stdout for other communication
+- Added option to output structured JSON logs instead of pretty colored terminal logs
 
 ## [v0.1.3]
 - Changed distribution archive to zip